move function to era package
